runtime/python/hailo_onnxruntime_example/requirements.txt
numpy
zenlog
Pillow>=8.1.2
onnxruntime
onnx
onnxruntime-openvino
